requirements.txt
streamlit
pandas
numpy
scikit-learn
langchain
langchain-openai
pypdf
faiss-cpu
tiktoken
python-dotenv
mcp
pydantic
tenacity
requests
langchain-google-genai
google-generativeai
pdfminer.six
matplotlib
seaborn
langchain-community
sentence-transformers
langchain-huggingface
fastapi
uvicorn